+++ src/Knet.jl
@@ -5,12 +5,11 @@ using Compat
 
 const libknet8 = Libdl.find_library(["libknet8"], [dirname(@__FILE__)])
 
-using AutoGrad; export grad, gradloss, gradcheck
-
+using AutoGrad; export grad, gradloss, gradcheck, getval
 
 include("gpu.jl");              export gpu
 include("kptr.jl");             # KnetPtr
-include("karray.jl");           export KnetArray, cpu2gpu, gpu2cpu
+include("karray.jl");           export KnetArray
 include("unary.jl");            export relu, sigm, invx, logp
 include("broadcast.jl");        # elementwise broadcasting operations
 include("reduction.jl");        export logsumexp