+++ test/karray.jl
@@ -6,7 +6,6 @@ else
 end
 
 using Knet
-import Base: (==)
 
 # Test KnetArray operations: cat, convert, copy, display, eachindex,
 # eltype, endof, fill!, first, getindex, hcat, isempty, length,
@@ -14,7 +13,6 @@ import Base: (==)
 # similar, size, stride, strides, summary, vcat, vec, zeros
 
 if gpu() >= 0
-    (==)(a::Array,k::KnetArray)=(a==Array(k))
     @testset "karray" begin
         a = rand(3,4)
         k = KnetArray(a)
@@ -74,12 +72,25 @@ if gpu() >= 0
             @test isa(pointer(k,3), Ptr{Float64})
             @test isempty(KnetArray(Float32,0))
             @test rand!(copy(a)) != rand!(copy(k))
+            @test k == k
+            @test a == k
+            @test k == a
+            @test isapprox(k,k)
+            @test isapprox(a,k)
+            @test isapprox(k,a)
+            @test a == copy!(similar(a),k)
+            @test k == copy!(similar(k),a)
+            @test k == copy!(similar(k),k)
+            @test k == copy(k)
+            @test pointer(k) != pointer(copy(k))
+            @test k == deepcopy(k)
+            @test pointer(k) != pointer(deepcopy(k))
         end
 
         @testset "cpu2gpu" begin
             # cpu/gpu xfer with grad support
-            @test gradcheck(x->gpu2cpu(sin(cpu2gpu(x))),a)
-            @test gradcheck(x->cpu2gpu(sin(gpu2cpu(x))),k)
+            @test gradcheck(x->Array(sin(KnetArray(x))),a)
+            @test gradcheck(x->KnetArray(sin(Array(x))),k)
         end
     end
 end