+++ test/runtests.jl
@@ -1,11 +1,11 @@
-#                                   tr  osx ai  ai4 tr4
-@time include("kptr.jl")          # 0   0   1   1   0
-@time include("gpu.jl")           # 0   0   1   1   0
-@time include("distributions.jl") # 2   3   1   1   1
-@time include("karray.jl")        # 0   0   9   5   0
-@time include("linalg.jl")        # 9   17  10  6   5
-@time include("update.jl")        # 77  18  19  17  19
-@time include("broadcast.jl")     # 21  50  19  10  7
-@time include("reduction.jl")     # 18  45  21  8   6
-@time include("unary.jl")         # 31  54  30  4   4
-@time include("conv.jl")          # 182 66  25  15  13
+#                                   ai  ai4 tr  tr4 aws osx 
+@time include("kptr.jl")          # 1   1   0   0   2   0   
+@time include("gpu.jl")           # 1   1   0   0   1   0   
+@time include("distributions.jl") # 1   1   2   1   2   3   
+@time include("karray.jl")        # 10  6   0   0   15  0   
+@time include("linalg.jl")        # 10  6   9   5   15  17  
+@time include("update.jl")        # 19  17  76  19  40  18  
+@time include("broadcast.jl")     # 19  10  19  7   29  50  
+@time include("reduction.jl")     # 21  8   16  6   31  45  
+@time include("unary.jl")         # 30  4   27  4   43  54  
+@time include("conv.jl")          # 25  15  57  13  37  66